backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
urllib3>=2.0.0
rank-bm25>=0.2.2
networkx>=3.0
faiss-cpu>=1.7.0
tiktoken>=0.5.0
langchain-core>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langsmith>=0.1.0
pytest>=7.0.0
aiosqlite>=0.19.0
bcrypt>=4.0.0
PyJWT>=2.8.0
